fix: クイック精算の負担詳細表示問題を解決 - ログイン済み版と同じ詳細な精算結果を出力
- showDetailedBreakdownをtrueに変更し負担詳細表示を有効化
- ID変換ロジックを改善（文字列ID→数値IDの適切な変換）
- NextAuth Prisma Adapterを@auth/prisma-adapterに移行
- authOptionsを独立したファイル(src/lib/auth.ts)に分離
- CSS最適化設定を無効化してDocker環境での安定動作を確保
