+++ package.json
@@ -16,9 +16,10 @@
     "db:generate": "prisma generate"
   },
   "dependencies": {
-    "@next-auth/prisma-adapter": "^1.0.7",
+    "@auth/prisma-adapter": "^2.10.0",
     "@prisma/client": "^5.7.1",
     "clsx": "^1.2.1",
+    "critters": "^0.0.23",
     "lucide-react": "^0.263.1",
     "next": "14.1.0",
     "next-auth": "^4.24.5",
